lib: bsdlib: return errno instead of EINVAL when unable to translate
When an unknown errno is received from bsdlib, bsd_os.c translates
the errno to EINVAL, effectively hiding it. This is confusing,
since EINVAL is a very generic errno that can be returned from
a myriad of different places in the code.

Return the original, non-translated errno which can at least be
used for debugging purposes and to more easily spot the case
when an errno is not being translated.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>